Added feature to avoid MetaEvents duplicity
As MIDI supports just one metadata per type (e. g. if it has two tempos assigned to the same track, only the last one will be applied), LuaMidi overrides MetaEvents of the same type, if they're set to the same track. It makes it possible to edit these events by setting them again (which overrides the last that was set with the one to be set).